mail-listener/requirements.txt
watchdog
kafka-python
emlx
html2text
